Consolidate tax code options
