+++ src/components/BankTransactionsMobileList/BankTransactionsMobileListSplitForm.tsx
@@ -25,7 +25,7 @@ import { AmountInput } from '@components/Input/AmountInput'
 import { FileInput } from '@components/Input/FileInput'
 import { Input } from '@components/Input/Input'
 import type { TaxCodeSelectOption } from '@components/TaxCodeSelect/constants'
-import { TaxCodeSelectDrawerWithTrigger } from '@components/TaxCodeSelect/TaxCodeSelectDrawerWithTrigger'
+import { TaxCodeSelect } from '@components/TaxCodeSelect/TaxCodeSelect'
 import { ErrorText } from '@components/Typography/ErrorText'
 
 import './bankTransactionsMobileListSplitForm.scss'
@@ -406,7 +406,8 @@ const SplitTaxCodeSelect = ({
   return (
     <>
       <HStack pis='3xs' aria-hidden='true' />
-      <TaxCodeSelectDrawerWithTrigger
+      <TaxCodeSelect
+        isMobileView
         options={taxCodeOptions}
         value={getSelectedTaxCodeOption(split.taxCode ?? null)}
         onChange={handleTaxCodeChange(splitIndex)}
